+++ CHANGES.rst
@@ -7,6 +7,8 @@ Version 0.2
 
 not yet released
 
+- Return nothing when parsing irrelevant HTML table rows.
+
 
 Version 0.1
 -----------